revolution-index/requirements.txt
fredapi>=0.5.1
pandas>=2.0
numpy>=1.24
scipy>=1.10
matplotlib>=3.7
seaborn>=0.12
plotly>=5.15
pyarrow>=12.0
requests>=2.31
python-dotenv>=1.0
jupyter>=1.0
pytest>=7.0
